test: move all uni-builder e2e cases to single folder
