+++ e2e/cases/css/css-modules/index.test.ts
@@ -1,4 +1,3 @@
-import path from 'path';
 import { expect, test } from '@playwright/test';
 import { build } from '@scripts/shared';
 
@@ -23,33 +22,6 @@ test('should compile CSS modules correctly', async () => {
   } else {
     expect(content).toEqual(
       '.the-a-class{color:red}.the-b-class-_HnKpz{color:blue}.the-c-class-e94QZl{color:#ff0}.the-d-class{color:green}',
-    );
-  }
-});
-
-// TODO: uni-builder
-test.skip('should treat normal CSS as CSS modules when disableCssModuleExtension is true', async () => {
-  const rsbuild = await build({
-    cwd: __dirname,
-    rsbuildConfig: {
-      output: {
-        disableSourceMap: true,
-        // disableCssModuleExtension: true,
-      },
-    },
-  });
-  const files = await rsbuild.unwrapOutputJSON();
-
-  const content =
-    files[Object.keys(files).find((file) => file.endsWith('.css'))!];
-
-  if (rsbuild.providerType === 'rspack') {
-    expect(content).toEqual(
-      '.the-a-class-_932a3{color:red}.the-b-class-_6773e{color:blue}.the-c-class-c855fd{color:yellow}.the-d-class{color:green}',
-    );
-  } else {
-    expect(content).toEqual(
-      '.the-a-class-azoWcU{color:red}.the-b-class-_HnKpz{color:blue}.the-c-class-e94QZl{color:#ff0}.the-d-class{color:green}',
     );
   }
 });
