+++ e2e/cases/html/title/index.test.ts
@@ -84,33 +84,3 @@ test('should generate title via function correctly', async () => {
     files[Object.keys(files).find((file) => file.endsWith('bar.html'))!];
   expect(barHtml).toContain('<title>bar</title>');
 });
-
-// TODO move to uni-builder
-test.skip('should generate title for MPA correctly', async () => {
-  const rsbuild = await build({
-    cwd: __dirname,
-    rsbuildConfig: {
-      source: {
-        entry: {
-          foo: path.resolve(__dirname, './src/foo.js'),
-          bar: path.resolve(__dirname, './src/foo.js'),
-        },
-      },
-      html: {
-        title: 'default',
-        // titleByEntries: {
-        //   foo: 'foo',
-        // },
-      },
-    },
-  });
-  const files = await rsbuild.unwrapOutputJSON();
-
-  const fooHtml =
-    files[Object.keys(files).find((file) => file.endsWith('foo.html'))!];
-  expect(fooHtml).toContain('<title>foo</title>');
-
-  const barHtml =
-    files[Object.keys(files).find((file) => file.endsWith('bar.html'))!];
-  expect(barHtml).toContain('<title>default</title>');
-});
