+++ e2e/uni-builder/css-modules/src/a.css
@@ -1,0 +1,3 @@
+.the-a-class {
+  color: red;
+}
