+++ e2e/uni-builder/css-modules/src/b.module.scss
@@ -1,0 +1,3 @@
+.the-b-class {
+  color: blue;
+}
