+++ e2e/uni-builder/css-modules/src/c.module.less
@@ -1,0 +1,3 @@
+.the-c-class {
+  color: yellow;
+}
