+++ e2e/uni-builder/css-modules/src/d.global.less
@@ -1,0 +1,3 @@
+.the-d-class {
+  color: green;
+}
