+++ e2e/uni-builder/css-modules/src/index.js
@@ -1,0 +1,4 @@
+import './a.css';
+import './b.module.scss';
+import './c.module.less';
+import './d.global.less';
