+++ e2e/uni-builder/global-vars/src/types.d.ts
@@ -1,0 +1,1 @@
+declare const ENABLE_TEST: boolean | undefined;
