+++ e2e/uni-builder/import-antd-v5/package.json
