+++ e2e/uni-builder/inline-chunk/src/foo.js
@@ -1,0 +1,4 @@
+export function foo() {
+  console.log('foo');
+  return 'foo';
+}
