+++ e2e/uni-builder/inline-chunk/src/index.js
@@ -1,0 +1,12 @@
+import './style.css';
+
+// eslint-disable-next-line no-undef
+window.test = 'aaaa';
+
+import(
+  /* webpackChunkName: "foo" */
+  './foo'
+).then(({ foo }) => {
+  // eslint-disable-next-line no-undef
+  window.answer = foo();
+});
