+++ e2e/uni-builder/inline-chunk/src/style.css
@@ -1,0 +1,5 @@
+html,
+body {
+  width: 100%;
+  height: 100%;
+}
