+++ e2e/uni-builder/manifest/index.test.ts
