+++ e2e/uni-builder/manifest/src/index.jsx
