+++ e2e/uni-builder/meta/src/foo.js
