+++ e2e/uni-builder/module-scopes/common/test.js
