+++ e2e/uni-builder/module-scopes/src/App.jsx
