+++ e2e/uni-builder/module-scopes/src/index.js
