+++ e2e/uni-builder/resolve-extension-prefix/src/ex.js
