+++ e2e/uni-builder/resolve-extension-prefix/src/index.js
