+++ e2e/uni-builder/resolve-extension-prefix/tests/index.test.ts
