+++ e2e/uni-builder/sri/index.test.ts
