+++ e2e/uni-builder/sri/src/index.js
