+++ e2e/uni-builder/title/index.test.ts
@@ -1,0 +1,32 @@
+import path from 'path';
+import { expect, test } from '@playwright/test';
+import { build } from '@scripts/shared';
+
+test.skip('should generate title for MPA correctly', async () => {
+  const rsbuild = await build({
+    cwd: __dirname,
+    rsbuildConfig: {
+      source: {
+        entry: {
+          foo: path.resolve(__dirname, './src/foo.js'),
+          bar: path.resolve(__dirname, './src/foo.js'),
+        },
+      },
+      html: {
+        title: 'default',
+        // titleByEntries: {
+        //   foo: 'foo',
+        // },
+      },
+    },
+  });
+  const files = await rsbuild.unwrapOutputJSON();
+
+  const fooHtml =
+    files[Object.keys(files).find((file) => file.endsWith('foo.html'))!];
+  expect(fooHtml).toContain('<title>foo</title>');
+
+  const barHtml =
+    files[Object.keys(files).find((file) => file.endsWith('bar.html'))!];
+  expect(barHtml).toContain('<title>default</title>');
+});
