+++ e2e/uni-builder/title/src/empty.html
@@ -1,0 +1,3 @@
+<!DOCTYPE html>
+<head></head>
+</html>
