+++ e2e/uni-builder/title/src/foo.js
@@ -1,0 +1,1 @@
+console.log('1');
